renamed contentClasses file 'classes'
